requirements.txt
absl-py==0.9.0
alabaster==0.7.11
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==19.3.0
Automat==0.7.0
autopep8==1.5
Babel==2.6.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.3
bencoder==0.2.0
bencoder.pyx==2.0.1
bitarray==0.8.3
bleach==2.1.4
boto==2.49.0
cachetools==4.0.0
certifi==2018.8.24
cffi==1.14.0
chardet==3.0.4
click==6.7
cloudpickle==0.5.5
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.7
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.8
cycler==0.10.0
Cython==0.28.5
cytoolz==0.9.0.1
dask==0.19.1
decorator==4.3.0
defusedxml==0.5.0
distributed==1.23.1
docutils==0.14
elasticsearch==5.1.0
elasticsearch-dsl >=5.0.0, <6.0.0
entrypoints==0.2.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
fastcache==1.0.2
filelock==3.0.8
Flask==1.0.2
Flask-Cors==3.0.6
gevent==1.3.6
glob2==0.6
google-auth==1.11.2
google-auth-oauthlib==0.4.1
greenlet==0.4.15
grpcio==1.27.2
heapdict==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
imagesize==1.1.0
incremental==17.5.0
ipykernel==4.10.0
ipython==6.5.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
keyring==13.2.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.24.0
locket==0.2.0
lxml==4.2.5
Markdown==3.2.1
MarkupSafe==1.0
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
more-itertools==4.3.0
mpmath==1.0.0
msgpack==0.5.6
multipledispatch==0.6.0
mysql-connector-python==8.0.18
mysqlclient==1.4.6
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.1
nltk==3.3
nose==1.3.7
numpy==1.15.1
oauthlib==3.1.0
olefile==0.46
openpyxl==2.5.6
packaging==17.1
pandocfilters==1.4.2
parsel==1.5.2
parso==0.3.1
partd==0.3.8
path.py==11.1.0
pathlib2==2.3.2
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.2.0
pkginfo==1.4.2
pluggy==0.7.1
ply==3.11
prompt-toolkit==1.0.15
Protego==0.1.16
protobuf==3.11.3
psutil==5.4.7
py==1.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
PyDispatcher==2.0.5
PyExecJS==1.5.1
pyflakes==2.0.0
Pygments==2.2.0
PyHamcrest==2.0.0
pylint==2.1.1
PyMySQL==0.9.3
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
pytesseract==0.3.2
pytest==3.8.0
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-dateutil==2.7.3
pytz==2018.5
pywin32==223
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.4.4
QtPy==1.5.0
queuelib==1.5.0
redis==3.4.1
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.11.0
rsa==4.0
ruamel-yaml==0.15.46
Scrapy==1.8.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.0.5
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.11
sympy==1.1.1
tblib==1.3.2
tensorboard==2.0.2
terminado==0.8.1
tesseract==0.1.3
testpath==0.3.1
toolz==0.9.0
tornado==5.1
tqdm==4.26.0
traitlets==4.3.2
Twisted==19.10.0
unicodecsv==0.14.1
urllib3==1.23
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.1.0
xlwings==0.11.8
xlwt==1.3.0
zict==0.1.3
zope.interface==4.5.0
